Refactor filtering code
Avoid duplicate sections, and emphasize split between retrieval/filtering as
well as tree/branch. The latter refer to different variations of a given kmer.
A tree involves changing the first and last base of a kmer, while a branch is
a subset of a tree and involves changing the last base only.