Turn values intro matrices of first/last variants
In order to improve locality during the filtering phase, values now contain
counters to all variants of a kmer, meaning all possible changes to the first
and last bases.

Each value is constructed as a matrix of uint32_t elements as follows:
 - 1st dimension: first base, can be any of [ACGT] encoded as [0123].
 - 2nd dimension: last base, can be any of [ACGT] encoded as [0123].
 - 3rd dimension: kind of read [NORMAL_READ, CANCER_READ].

Since each entry now contains counters for 32 kmers (4x4x2), instead of
hashing the whole kmer to index the values, the first and last bases are
simply skipped.